refactor: 도메인에 있는 출력 로직을 뷰로 이동
